Change dog hunger behaviour
Modfier instead of max cap  in config
Refactor
halves the speed when the dog is not moving.

+++ src/api/java/doggytalents/api/inferface/IDogAlteration.java
@@ -81,7 +81,7 @@ public interface IDogAlteration {
 
     }
 
-    default InteractionResultHolder<Integer> hungerTick(AbstractDog dogIn, int hungerTick) {
+    default InteractionResultHolder<Float> hungerTick(AbstractDog dogIn, float hungerTick) {
         return InteractionResultHolder.pass(hungerTick);
     }
